Checklist Completed

+++ resources/views/users/index.blade.php
@@ -34,10 +34,17 @@
                                 </td>
                                 <td>
                                     <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-                                    <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-                                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                                    {!! Form::close() !!}
+                                    @if(Auth::user()->hasRole('BBA'))
+                                        @if( $user->email != 'admin@blackbox.com' )
+                                        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                                        {!! Form::close() !!}
+                                        @endif
+                                    @endif
+                                    @if(Auth::user()->hasRole(['BBA','TL','Recruiter']))
+                                        <a class="btn btn-success" href="{{ route('users.checklists',$user->id) }}">Checklist</a>
+                                    @endif
                                 </td>
                             </tr>
                         @endforeach
